+++ src/device_impl.rs
@@ -3,7 +3,11 @@ use crate::{
     calculate_raw_threshold_value, Config, Error, FaultCount, Gain, IntegrationTime,
     InterruptStatus, PowerSavingMode, SlaveAddr, Veml6030,
 };
-use embedded_hal::blocking::i2c;
+
+#[cfg(not(feature = "async"))]
+use embedded_hal::i2c::I2c;
+#[cfg(feature = "async")]
+use embedded_hal_async::i2c::I2c as AsyncI2c;
 
 struct Register;
 impl Register {
@@ -58,28 +62,37 @@ impl<I2C> Veml6030<I2C> {
     }
 }
 
+#[maybe_async_cfg::maybe(
+    sync(
+        cfg(not(feature = "async")),
+        self = "Veml6030",
+        idents(AsyncI2c(sync = "I2c"))
+    ),
+    async(feature = "async", keep_self)
+)]
+
 impl<I2C, E> Veml6030<I2C>
 where
-    I2C: i2c::Write<Error = E>,
+    I2C: AsyncI2c<Error = E>,
 {
     /// Enable the device.
     ///
     /// Note that when activating the sensor a wait time of 4 ms should be
     /// observed before the first measurement is picked up to allow for a
     /// correct start of the signal processor and oscillator.
-    pub fn enable(&mut self) -> Result<(), Error<E>> {
+    pub async fn enable(&mut self) -> Result<(), Error<E>> {
         let config = self.config.with_low(BitFlags::ALS_SD);
-        self.set_config(config)
+        self.set_config(config).await
     }
 
     /// Disable the device (shutdown).
-    pub fn disable(&mut self) -> Result<(), Error<E>> {
+    pub async fn disable(&mut self) -> Result<(), Error<E>> {
         let config = self.config.with_high(BitFlags::ALS_SD);
-        self.set_config(config)
+        self.set_config(config).await
     }
 
     /// Set the integration time.
-    pub fn set_integration_time(&mut self, it: IntegrationTime) -> Result<(), Error<E>> {
+    pub async fn set_integration_time(&mut self, it: IntegrationTime) -> Result<(), Error<E>> {
         let mask = match it {
             IntegrationTime::Ms25 => 0b1100,
             IntegrationTime::Ms50 => 0b1000,
@@ -89,13 +102,13 @@ where
             IntegrationTime::Ms800 => 0b0011,
         };
         let config = self.config.bits & !(0b1111 << 6) | (mask << 6);
-        self.set_config(Config { bits: config })?;
+        self.set_config(Config { bits: config }).await?;
         self.it = it;
         Ok(())
     }
 
     /// Set the gain.
-    pub fn set_gain(&mut self, gain: Gain) -> Result<(), Error<E>> {
+    pub async fn set_gain(&mut self, gain: Gain) -> Result<(), Error<E>> {
         let mask = match gain {
             Gain::One => 0,
             Gain::Two => 1,
@@ -103,14 +116,14 @@ where
             Gain::OneQuarter => 3,
         };
         let config = self.config.bits & !(0b11 << 11) | (mask << 11);
-        self.set_config(Config { bits: config })?;
+        self.set_config(Config { bits: config }).await?;
         self.gain = gain;
         Ok(())
     }
 
     /// Set the number of times a threshold crossing must happen consecutively
     /// to trigger an interrupt.
-    pub fn set_fault_count(&mut self, fc: FaultCount) -> Result<(), Error<E>> {
+    pub async fn set_fault_count(&mut self, fc: FaultCount) -> Result<(), Error<E>> {
         let mask = match fc {
             FaultCount::One => 0,
             FaultCount::Two => 1,
@@ -118,29 +131,29 @@ where
             FaultCount::Eight => 3,
         };
         let config = self.config.bits & !(0b11 << 4) | (mask << 4);
-        self.set_config(Config { bits: config })
+        self.set_config(Config { bits: config }).await
     }
 
     /// Enable interrupt generation.
-    pub fn enable_interrupts(&mut self) -> Result<(), Error<E>> {
+    pub async fn enable_interrupts(&mut self) -> Result<(), Error<E>> {
         let config = self.config.with_high(BitFlags::ALS_INT_EN);
-        self.set_config(config)
+        self.set_config(config).await
     }
 
     /// Disable interrupt generation.
-    pub fn disable_interrupts(&mut self) -> Result<(), Error<E>> {
+    pub async fn disable_interrupts(&mut self) -> Result<(), Error<E>> {
         let config = self.config.with_low(BitFlags::ALS_INT_EN);
-        self.set_config(config)
+        self.set_config(config).await
     }
 
     /// Set the ALS high threshold in raw format
-    pub fn set_high_threshold_raw(&mut self, threshold: u16) -> Result<(), Error<E>> {
-        self.write_register(Register::ALS_WH, threshold)
+    pub async fn set_high_threshold_raw(&mut self, threshold: u16) -> Result<(), Error<E>> {
+        self.write_register(Register::ALS_WH, threshold).await
     }
 
     /// Set the ALS low threshold in raw format
-    pub fn set_low_threshold_raw(&mut self, threshold: u16) -> Result<(), Error<E>> {
-        self.write_register(Register::ALS_WL, threshold)
+    pub async fn set_low_threshold_raw(&mut self, threshold: u16) -> Result<(), Error<E>> {
+        self.write_register(Register::ALS_WL, threshold).await
     }
 
     /// Set the ALS high threshold in lux.
@@ -148,9 +161,9 @@ where
     /// For values higher than 1000 lx and 1/4 or 1/8 gain,
     /// the inverse of the compensation formula is applied (this involves
     /// quite some math).
-    pub fn set_high_threshold_lux(&mut self, lux: f32) -> Result<(), Error<E>> {
+    pub async fn set_high_threshold_lux(&mut self, lux: f32) -> Result<(), Error<E>> {
         let raw = self.calculate_raw_threshold_value(lux);
-        self.write_register(Register::ALS_WH, raw)
+        self.write_register(Register::ALS_WH, raw).await
     }
 
     /// Set the ALS low threshold in lux.
@@ -158,9 +171,9 @@ where
     /// For values higher than 1000 lx and 1/4 or 1/8 gain,
     /// the inverse of the compensation formula is applied (this involves
     /// quite some math).
-    pub fn set_low_threshold_lux(&mut self, lux: f32) -> Result<(), Error<E>> {
+    pub async fn set_low_threshold_lux(&mut self, lux: f32) -> Result<(), Error<E>> {
         let raw = self.calculate_raw_threshold_value(lux);
-        self.write_register(Register::ALS_WL, raw)
+        self.write_register(Register::ALS_WL, raw).await
     }
 
     /// Calculate raw value for threshold applying compensation if necessary.
@@ -176,7 +189,7 @@ where
     }
 
     /// Enable the power-saving mode
-    pub fn enable_power_saving(&mut self, psm: PowerSavingMode) -> Result<(), Error<E>> {
+    pub async fn enable_power_saving(&mut self, psm: PowerSavingMode) -> Result<(), Error<E>> {
         let mask = match psm {
             PowerSavingMode::One => 0,
             PowerSavingMode::Two => 1,
@@ -184,38 +197,34 @@ where
             PowerSavingMode::Four => 3,
         };
         let value = BitFlags::PSM_EN | (mask << 1);
-        self.write_register(Register::PSM, value)
+        self.write_register(Register::PSM, value).await
     }
 
     /// Disable the power-saving mode
-    pub fn disable_power_saving(&mut self) -> Result<(), Error<E>> {
-        self.write_register(Register::PSM, 0)
+    pub async fn disable_power_saving(&mut self) -> Result<(), Error<E>> {
+        self.write_register(Register::PSM, 0).await
     }
 
-    fn set_config(&mut self, config: Config) -> Result<(), Error<E>> {
-        self.write_register(Register::ALS_CONF, config.bits)?;
+    async fn set_config(&mut self, config: Config) -> Result<(), Error<E>> {
+        self.write_register(Register::ALS_CONF, config.bits).await?;
         self.config = config;
         Ok(())
     }
 
-    fn write_register(&mut self, register: u8, value: u16) -> Result<(), Error<E>> {
+    async fn write_register(&mut self, register: u8, value: u16) -> Result<(), Error<E>> {
         self.i2c
             .write(self.address, &[register, value as u8, (value >> 8) as u8])
+            .await
             .map_err(Error::I2C)
     }
-}
 
-impl<I2C, E> Veml6030<I2C>
-where
-    I2C: i2c::WriteRead<Error = E>,
-{
     /// Read whether an interrupt has occurred.
     ///
     /// Note that the interrupt status is updated at the same rate as the
     /// measurements. Once triggered, flags will stay true until a measurement
     /// is taken which does not exceed the threshold.
-    pub fn read_interrupt_status(&mut self) -> Result<InterruptStatus, Error<E>> {
-        let data = self.read_register(Register::ALS_INT)?;
+    pub async fn read_interrupt_status(&mut self) -> Result<InterruptStatus, Error<E>> {
+        let data = self.read_register(Register::ALS_INT).await?;
         Ok(InterruptStatus {
             was_too_low: (data & BitFlags::INT_TH_LOW) != 0,
             was_too_high: (data & BitFlags::INT_TH_HIGH) != 0,
@@ -223,8 +232,8 @@ where
     }
 
     /// Read ALS high resolution output data in raw format
-    pub fn read_raw(&mut self) -> Result<u16, Error<E>> {
-        self.read_register(Register::ALS)
+    pub async fn read_raw(&mut self) -> Result<u16, Error<E>> {
+        self.read_register(Register::ALS).await
     }
 
     /// Read ALS high resolution output data converted to lux
@@ -232,8 +241,8 @@ where
     /// For values higher than 1000 lx and 1/4 or 1/8 gain,
     /// the following compensation formula is applied:
     /// `lux = 6.0135e-13*(lux^4) - 9.3924e-9*(lux^3) + 8.1488e-5*(lux^2) + 1.0023*lux`
-    pub fn read_lux(&mut self) -> Result<f32, Error<E>> {
-        let raw = self.read_register(Register::ALS)?;
+    pub async fn read_lux(&mut self) -> Result<f32, Error<E>> {
+        let raw = self.read_register(Register::ALS).await?;
         Ok(self.convert_raw_als_to_lux(raw))
     }
 
@@ -250,14 +259,15 @@ where
     }
 
     /// Read white channel measurement
-    pub fn read_white(&mut self) -> Result<u16, Error<E>> {
-        self.read_register(Register::WHITE)
+    pub async fn read_white(&mut self) -> Result<u16, Error<E>> {
+        self.read_register(Register::WHITE).await
     }
 
-    fn read_register(&mut self, register: u8) -> Result<u16, Error<E>> {
+    async fn read_register(&mut self, register: u8) -> Result<u16, Error<E>> {
         let mut data = [0; 2];
         self.i2c
             .write_read(self.address, &[register], &mut data)
+            .await
             .map_err(Error::I2C)
             .and(Ok(u16::from(data[0]) | (u16::from(data[1]) << 8)))
     }